+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixes
 
 * Fix completion when update is available
+* Fix repack when there is composer dependencies to castor
 
 ### Features